Added image and closed #15
In regards to #15, it seems that adding the 429 retry to the prometheus
config works well enough. I ran this for over 12 hours on my linux
machine and got many 429 errors, but the retry seemed to always succeed.
Even though I got so many 429 errors. I would love to figure out why
this occurs and how to correctly avoid getting it in the first place,
but I wasn't able to find any sufficient resources on this topic.